Move artifacts definition to separate file
